requirements.txt
Django==1.8.5
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
boto==2.38.0
celery==3.1.19
dj-database-url==0.3.0
django-appconf==1.0.1
django-ckeditor==5.0.2
django-geoposition==0.2.2
django-redis-cache==1.6.3
django-s3direct==0.3.11
django-safedelete==0.3.1
django-storages-redux==1.3
gunicorn==19.3.0
hiredis==0.2.0
kombu==3.0.29
psycopg2==2.6.1
pytz==2015.7
raven==5.8.1
redis==2.10.5
six==1.10.0
whitenoise==2.0.4
